+++ CHANGELOG.md
@@ -15,7 +15,7 @@
   * Remove bond (outdated)
   * Leave binding_of_caller optional
 
-* Add command wrappers for looksee + shadow
+* Add command wrappers for looksee, shadow, howtocall
 * Add `$` command to directly call system
 
 * Version bumps: looksee, os, methodfinder, code, core_docs, debugging